Add files via upload
What is Data Extraction?
Data extraction is the initial phase of the ETL process. This involves pulling data from a range of source systems, including but not limited to databases, CRM systems, or flat files like CSVs. The objective is to collect as much relevant data as possible to power our analytics systems.

The Transformation Process
Once the data is extracted, it enters the transformation stage. Here, the data is cleaned, validated, and prepared to meet the requirements of the business or the target system. This could involve removing duplicates, correcting errors, or converting data into compatible formats. The goal is to maintain data integrity and quality before it is loaded into the data warehouse.

Loading Data
After the data has been transformed, it is then loaded into a data warehouse or other target system. This is where the processed data will reside for further use, usually for analytics, business intelligence, or reporting purposes.

ETL in Python
Python is a great language for performing ETL because of its readability and the wide range of data-centric libraries it offers. Here, we’re going to use Python to perform ETL on datasets from kaggle https://kaggle.com/datasets/andrewmvd/data-scientist-jobs/data . We’ll use the Pandas and NumPy libraries for data manipulation.